# go-server-server-generated/go/model_receipt.go
/*
 * Receipt Processor
 *
 * A simple receipt processor
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type Receipt struct {
	// The name of the retailer or store the receipt is from.
	Retailer string `json:"retailer"`
	// The date of the purchase printed on the receipt.
	PurchaseDate string `json:"purchaseDate"`
	// The time of the purchase printed on the receipt. 24-hour time expected.
	PurchaseTime string `json:"purchaseTime"`

	Items []Item `json:"items"`
	// The total amount paid on the receipt.
	Total string `json:"total"`
}